requirements.txt
fabric
requests
tornado
pyyaml
crontab
pycrypto
